refactor: RegisterForm
