--- src/components/elements/Forms/styles.module.scss
@@ -1,46 +1,0 @@
-.errors {
-  color: red;
-  font-size: 14px;
-  padding: 0;
-  margin: 0;
-}
-
-.form {
-  margin: 0 auto;
-  display: flex;
-  flex-direction: column;
-  max-width: 320px;
-  & label {
-    font-size: 14px;
-    line-height: 20px;
-    font-weight: 700;
-  }
-}
-
-.checkbox_block {
-  display: flex;
-  flex-direction: column;
-  & p {
-    font-size: 16px;
-    line-height: 18px;
-  }
-}
-
-.label_checkbox {
-  display: flex;
-  align-items: baseline;
-  p {
-    margin-left: 15px;
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 18px;
-    & a {
-      cursor: pointer;
-    }
-  }
-}
-.input {
-  margin: 0;
-  padding: 0;
-  border: 1px solid red;
-}
